Fixed data type errors

+++ plugins/wasi_nn/wasinn_ggml.cpp
@@ -509,8 +509,8 @@ ErrNo evaluateTokens(Graph &GraphRef, struct llama_context *LlamaContext,
                      int &NPos) noexcept {
   uint32_t NCtx = llama_n_ctx(LlamaContext);
 
-  if (NPast > std::numeric_limits<uint32_t>::max() -
-                  static_cast<uint32_t>(Tokens.size())) {
+  if (static_cast<uint32_t>(NPast) > std::numeric_limits<uint32_t>::max() -
+                                         static_cast<uint32_t>(Tokens.size())) {
     spdlog::error(
         "[WASI-NN] GGML backend: integer overflow detected when calculating context size."sv);
     return ErrNo::RuntimeError;
